Atualização do BACKEND

+++ MyIce-BackEnd/src/main/java/ecommerce/Model/Pedido.java
@@ -26,7 +26,7 @@ public class Pedido {
     private Date dtPedido;
 
     @Column(name = "id_cliente")
-    private Long cliente;
+    private Integer cliente;
 
     @Column(name = "vl_frete")
     private BigDecimal vlFrete;
